requirements.txt
@"
streamlit==1.28.0
openai==1.35.0
python-docx==0.8.11
PyPDF2==3.0.1
python-dotenv==1.0.0
requests==2.31.0
Pillow==10.0.0
"@ | Out-File -FilePath "requirements.txt" -Encoding UTF8